requirements.txt
python-dotenv
Django==5.2.6
psycopg2-binary==2.9.10
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.9.1
flake8==7.3.0
isort==6.0.1
mypy==1.18.2
black==25.1.0